requirements.txt
flask
requests
deep-translator
python-dotenv
supabase